github_bot.py
"""
QuantumFlow Elite Bybit Bot - Main Bot File
Last Updated: 2025-03-29 12:35:04 UTC
Author: chibueze2345
Version: 2.1.0
"""

import os
import sys
import json
import logging
import time
from datetime import datetime, timezone
import asyncio
from telegram.ext import Application, CommandHandler
import telegram
from config import get_current_config
from storage import BotStorage

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('logs/bot.log'),
        logging.StreamHandler(sys.stdout)
    ]
)
logger = logging.getLogger(__name__)

class QuantumFlowBot:
    def __init__(self):
        self.config = get_current_config()
        self.is_testnet = self.config['environment'] == 'TESTNET'
        self.storage = BotStorage()
        
        # Set API credentials based on environment
        if self.is_testnet:
            self.api_key = os.environ['BYBIT_TESTNET_API_KEY']
            self.api_secret = os.environ['BYBIT_TESTNET_API_SECRET']
            self.base_url = self.config['api']['base_url']
        else:
            self.api_key = os.environ['BYBIT_API_KEY']
            self.api_secret = os.environ['BYBIT_API_SECRET']
            self.base_url = self.config['api']['base_url']

        self.telegram_token = os.environ['TELEGRAM_BOT_TOKEN']
        self.telegram_chat_id = os.environ['TELEGRAM_CHAT_ID']
        
        # Initialize state
        self.state_file = f"bot_state_{'testnet' if self.is_testnet else 'mainnet'}.json"
        self.load_state()
        
        # Initialize Telegram
        self.telegram = telegram.Bot(token=self.telegram_token)
        
    async def send_telegram_message(self, message):
        try:
            env_prefix = "🧪 TESTNET: " if self.is_testnet else "🚀 LIVE: "
            await self.telegram.send_message(
                chat_id=self.telegram_chat_id,
                text=f"{env_prefix}{message}",
                parse_mode='Markdown'
            )
        except Exception as e:
            logger.error(f"Failed to send Telegram message: {str(e)}")

    def load_state(self):
        try:
            stored_state = self.storage.load_data(self.state_file)
            if stored_state:
                self.state = stored_state
            else:
                self.state = {
                    'environment': self.config['environment'],
                    'last_run': None,
                    'positions': {},
                    'daily_trades': 0,
                    'balance': 0,
                    'pnl': 0
                }
        except Exception as e:
            logger.error(f"Failed to load state: {str(e)}")
            self.state = {}

    def save_state(self):
        try:
            self.storage.save_data(self.state, self.state_file)
        except Exception as e:
            logger.error(f"Failed to save state: {str(e)}")

    async def execute_trade(self, pair, direction, amount):
        try:
            if self.is_testnet:
                amount = min(amount, self.state['balance'] * 0.1)
            
            trade_result = await self._place_order(pair, direction, amount)
            
            await self.send_telegram_message(
                f"Trade executed:\n"
                f"Pair: {pair}\n"
                f"Direction: {direction}\n"
                f"Amount: {amount}\n"
                f"Result: {trade_result}"
            )
            
            return trade_result
            
        except Exception as e:
            logger.error(f"Trade execution error: {str(e)}")
            await self.send_telegram_message(f"❌ Trade execution failed: {str(e)}")
            return None

    async def run_cycle(self):
        try:
            current_time = datetime.now(timezone.utc)
            
            await self.send_telegram_message(
                f"🔄 Starting trading cycle\n"
                f"Time: {current_time.strftime('%Y-%m-%d %H:%M:%S')} UTC"
            )
            
            await self.update_market_data()
            await self.check_positions()
            await self.execute_strategy()
            
            self.save_state()
            
            await self.send_status_update()
            
        except Exception as e:
            logger.error(f"Error in run cycle: {str(e)}")
            await self.send_telegram_message(f"⚠️ Error in trading cycle: {str(e)}")

    async def update_market_data(self):
        # Implement market data update logic
        pass

    async def check_positions(self):
        # Implement position checking logic
        pass

    async def execute_strategy(self):
        # Implement trading strategy logic
        pass

    async def send_status_update(self):
        status = {
            'time': datetime.now(timezone.utc).strftime('%Y-%m-%d %H:%M:%S'),
            'balance': self.state.get('balance', 0),
            'positions': len(self.state.get('positions', {})),
            'daily_trades': self.state.get('daily_trades', 0),
            'pnl': self.state.get('pnl', 0)
        }
        
        await self.send_telegram_message(
            f"📊 *Status Update*\n\n"
            f"🕒 Time: {status['time']} UTC\n"
            f"💰 Balance: ${status['balance']:.2f}\n"
            f"📈 Open Positions: {status['positions']}\n"
            f"🎯 Today's Trades: {status['daily_trades']}\n"
            f"📗 P/L: ${status['pnl']:.2f}"
        )

async def main():
    bot = QuantumFlowBot()
    await bot.run_cycle()

if __name__ == "__main__":
    asyncio.run(main())