requirements.txt
# API and Trading
python-bybit==3.0.0
ccxt>=2.0.0

# Data Processing
pandas>=1.3.0
numpy>=1.21.0

# Telegram Integration
python-telegram-bot>=20.0
aiohttp>=3.8.0

# Utilities
python-dotenv>=0.19.0
schedule>=1.1.0
pytz>=2021.3
requests>=2.26.0

# Security
cryptography>=3.4.7
pyjwt>=2.1.0